feat: show consent modal for users that have a saved wallet
