+++ cypress/e2e/login_spec.ts
@@ -15,10 +15,8 @@ describe('The Dashboard', () => {
   it('supports log in via an imported Native wallet', () => {
     cy.visit('')
 
-    cy.clearLocalStorage()
     // Open WalletProvider.SelectModal
     cy.getBySel('connect-wallet-button').click()
-    cy.getBySel('consent-optin-continue-button').click()
     cy.getBySel('connect-wallet-native-button').click()
     cy.getBySel('wallet-native-import-button').click()
 
@@ -60,9 +58,7 @@ describe('The Dashboard', () => {
   it('supports login via locally stored Native wallet', () => {
     // This will use the wallet created in `supports log in via an imported Native wallet`
     cy.visit('')
-    cy.clearLocalStorage()
     cy.getBySel('connect-wallet-button').click()
-    cy.getBySel('consent-optin-continue-button').click()
     cy.getBySel('connect-wallet-native-button').click()
     cy.getBySel('wallet-native-load-button').click()
     cy.getBySel('native-saved-wallet').should('have.length', 1)
@@ -76,9 +72,7 @@ describe('The Dashboard', () => {
   it('cannot login natively when no local Native wallets', () => {
     cy.clearIndexedDB().then(() => {
       cy.visit('')
-      cy.clearLocalStorage()
       cy.getBySel('connect-wallet-button').click()
-      cy.getBySel('consent-optin-continue-button').click()
       cy.getBySel('connect-wallet-native-button').click()
       cy.getBySel('wallet-native-load-button').should('be.disabled')
     })
@@ -86,10 +80,8 @@ describe('The Dashboard', () => {
 
   it('support Portis log in', () => {
     cy.visit('')
-    cy.clearLocalStorage()
     // Open WalletProvider.SelectModal
     cy.getBySel('connect-wallet-button').click()
-    cy.getBySel('consent-optin-continue-button').click()
     cy.getBySel('connect-wallet-portis-button').click()
 
     // This would open an external page to log in with Portis, which is outside the scope of Cypress
