+++ src/context/WalletProvider/WalletViewsSwitch.tsx
@@ -9,7 +9,7 @@ import {
 } from '@chakra-ui/react'
 import { useToast } from '@chakra-ui/toast'
 import { AnimatePresence } from 'framer-motion'
-import { OptInModal } from 'plugins/pendo/components/OptInModal/OptInModal'
+import { OptInModalBody } from 'plugins/pendo/components/OptInModal/OptInModalBody'
 import { useCallback, useEffect } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { Route, Switch, useHistory, useLocation, useRouteMatch } from 'react-router-dom'
@@ -107,7 +107,10 @@ export const WalletViewsSwitch = () => {
                   })}
 
                 <Route path={'/select'} children={() => <SelectModal />} />
-                <Route path={'/'} children={() => <OptInModal />} />
+                <Route
+                  path={'/'}
+                  children={() => <OptInModalBody onContinue={() => history.push('/select')} />}
+                />
               </Switch>
             </SlideTransition>
           </AnimatePresence>
