--- src/plugins/analytics/components/OptInModal/OptInModal.stories.tsx
@@ -1,13 +1,0 @@
-/* eslint-disable import/no-anonymous-default-export */
-/* eslint-disable import/no-default-export */
-
-import { Story } from '@storybook/react'
-
-import { OptInModal } from './OptInModal'
-
-export default {
-  title: 'Plugins/Analytics/OptInModal',
-  component: OptInModal,
-}
-
-export const AnalyticsModal: Story = () => <OptInModal isOpen={true} close={() => {}} />
