+++ src/plugins/index.ts
@@ -1,10 +1,7 @@
 import { ChainId } from '@shapeshiftoss/caip'
 import { ChainAdapter } from '@shapeshiftoss/chain-adapters'
+import { Route } from 'Routes/helpers'
 import { FeatureFlags } from 'state/slices/preferencesSlice/preferencesSlice'
-
-import { Route } from '../Routes/helpers'
-
-const activePlugins = ['bitcoin', 'cosmos', 'ethereum', 'osmosis', 'avalanche']
 
 export type Plugins = [chainId: string, chain: Plugin][]
 export type RegistrablePlugin = { register: () => Plugins }
@@ -38,17 +35,5 @@ export class PluginManager {
 
   entries(): [string, Plugin][] {
     return [...this.#pluginManager.entries()]
-  }
-}
-
-// @TODO - In the future we may want to create a Provider for this
-// if we need to support features that require re-rendering. Currently we do not.
-export const pluginManager = new PluginManager()
-
-export const registerPlugins = async () => {
-  pluginManager.clear()
-
-  for (const plugin of activePlugins) {
-    pluginManager.register(await import(`./${plugin}/index.tsx`))
   }
 }
